mysql_with_config/requirements.txt
configparser==5.0.1
mysql-connector-python==8.0.22